Configuro tablas usuarios y funcionarios con vista

+++ resources/views/usuario/edit.blade.php
@@ -6,13 +6,27 @@
 
 @section('contenido')
 	@include('alerts.request')
-	<h1>Editar</h1>
-	{!!Form::model($user, ['route'=>['usuario.update', $user->id], 'method'=>'put'])!!}
-		@include('usuario.forms.user')
-		{!!Form::submit('Actualizar', ['class'=>'btn btn-primary'])!!}
-	{!!Form::close()!!}
+	@if(isset($user))
+		<h1>Editar</h1>
+		@if($user->idrol === 1)		
+			{!!Form::model($user, ['route'=>['usuario.update', $user->id], 'method'=>'put'])!!}
+				@include('usuario.forms.docente')
+				{!!Form::submit('Actualizar', ['class'=>'btn btn-primary'])!!}
+			{!!Form::close()!!}
 
-	{!!Form::open(['route'=>['usuario.destroy', $user->id], 'method'=>'delete'])!!}
-		{!!Form::submit('Eliminar', ['class'=>'btn btn-danger'])!!}
-	{!!Form::close()!!}
+			{!!Form::open(['route'=>['usuario.destroy', $user->id], 'method'=>'delete'])!!}
+				{!!Form::submit('Eliminar', ['class'=>'btn btn-danger'])!!}
+			{!!Form::close()!!}
+		@elseif($user->idrol === 2)
+			{!!Form::model($user, ['route'=>['usuario.update', $user->id], 'method'=>'put'])!!}
+				@include('usuario.forms.funcionario')
+				{!!Form::submit('Actualizar', ['class'=>'btn btn-primary'])!!}
+			{!!Form::close()!!}
+
+			{!!Form::open(['route'=>['usuario.destroy', $user->id], 'method'=>'delete'])!!}
+				{!!Form::submit('Eliminar', ['class'=>'btn btn-danger'])!!}
+			{!!Form::close()!!}
+		@endif
+
+	@endif
 @stop
